feat(stable-hash, transport): support `workerd`
