+++ stable-hash/dom.ts
